Organize main file
